+++ src/core/power/power.js
